requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
greenlet==3.0.3
idna==3.6
numpy==1.26.4
pandas==2.2.0
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
